+++ lib/shopify_app/engine.rb
@@ -16,10 +16,6 @@ module ShopifyApp
     engine_name "shopify_app"
     isolate_namespace ShopifyApp
 
-    initializer "shopify_app.middleware" do |app|
-      app.config.middleware.insert_after(::Rack::Runtime, ShopifyApp::JWTMiddleware)
-    end
-
     initializer "shopify_app.assets.precompile" do |app|
       app.config.assets.precompile += [
         "shopify_app/redirect.js",